Overwrites
The "append" operator (>>) to add content to the file specified by the variable %HISTORY_FILE%, while the second piece of code uses the "overwrite" operator (>) to write content to the same file, but discarding any existing content in it. The only difference between these two is that the first one adds to the content of the file, while the second one overwrites the content of the file.